+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to `antidot-fw/container` will be documented in this file.
 
 Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) principles.
 
+## 0.1.0 - 2020-05-31
+
+### Added
+
+* Docs site
+* Crete new Builder class with more easy config
+* Marc as deprecate ContainerBuilder class for config simplification
+
 ## 0.0.1 - 2019-09-16
 
 ### Added